requirements.txt
langchain==0.2.3langchain-chroma==0.1.3chromadb==0.4.24python-dotenvpyyamlrequests